Add a GitHub workflow to create Windows and Linux binaries
The Build workflow builds Debug and Release builds
for Windows and Linux and provides the binaries as
build artifacts.

The workflow uses the jurplel/install-qt-action to
fetch and install the Qt libraries. These are cached for
subsequent builds. This is used to install Ninja on
Linux and CMake, Ninja and MinGW on Windows.

The Windows build uses the Qt provided MinGW
toolchain rather than using MSYS from the GitHub
runner or installing separately as this seems to
consistently produce working binaries.

The Windows binaries have the require Qt
redistributables needed to run without installation of
anything further.

Tests:
 - Extensive testing on an experimental repo of both
    Windows and Linux binaries
 - Further testing after push