Log sim parameters and output
Create a JSON "log" of the parameters
input to the simulation and the output
of the simulation each time it is run. This
will allow automated tests to be written
in future and simple manual regression
tests in the short term.

The log format is a JSON array with a
parameters dictionary of the UI controls
at the start of the run. This is followed by
a results dictionary with arrays of data
points for each of the graphs.

The logging can be enabled or disabled
at compile time with the EnableLogging
global bool in mainwindow.cpp.

Log files start at 12MB and only get
bigger.

Tests:
 - Enable logging and verify syntax with
    Firefox.
 - Spot check data values looks sensible
    for both sim warm up runs and actual
    user input.
 - Verify disabling logging doesn't
    produce any output.