File log corrupt on Linux and enable by default
The log files are being corrupted on Linux because of
the log file gets trimmed by 1 extra character. On
Windows this is OK because the log files use cr/lf line-
endings. The fix is to measure the characters to
remove before rewinding the correct amount.

Enable logging by default as it is useful for on-going
regression testing prior to adding proper unit tests.

Tests:
 - Generate a series of log files and compare with
    golden master logs from Windows