Don't store motor params in MainWindow as well
Remove the shadow copies of vehicle parameters
used predominantly by MotorModel. Other than
m_Poles the values stored are never used until they
are replaced with a new value. Just transfer the values
straight from the UI into the model.

The Poles parameter on the MotorModel has a getter
added to allow it to be accessed during simulation.

Tests:
 - Run the Accel/Coast test against golden master logs
 - Vary a collection of parameters during a simulation
    run and observe the results change in a seemly valid
    way.