Use menu items to control simulation
Create a Simulation menu on the menu bar with
items for each of the existing options for
controlling the simulation.

The implementation of each control is delegated
to the MainWindow class which remains in charge
of running the simulation.

Use std::chrono::duration to represent the
variable time steps. Don't push this into the
simulation yet.

Tests:
 - Run each menu item and verify the output
   changes as expected.